omochi-be-master@1170e0e6cac/pyproject.toml
[project]
name = "omochi-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.2",
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "markdown>=3.8",
    "pillow>=11.2.1",
    "pip>=25.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "python-json-logger>=3.3.0",
    "qrcode>=8.1",
    "werkzeug>=3.1.3",
    "django-extensions",
    "stripe>=12.1.0",
    "openpyxl>=3.1.2",
    "firebase-admin>=6.9.0",
    "pytz>=2025.2",
    "django-soft-delete>=1.0.18",
    "openai>=1.58.1",
]
[tool.setuptools]
py-modules = []
